speed is now much improved after adding time.perf_counter() instead of time.time()
