--- telegrip/__init__.py
@@ -1,11 +1,0 @@
-"""
-TeleGrip - SO100 robot teleoperation system.
-"""
-
-from .config import TelegripConfig, load_config
-from .control_loop import ControlLoop
-from .core.robot_interface import RobotInterface
-from .core.visualizer import PyBulletVisualizer as Visualizer
-
-__version__ = "0.2.0"
-__all__ = ["RobotInterface", "Visualizer", "ControlLoop", "TelegripConfig", "load_config"]
