+++ telegrip/core/__init__.py
@@ -4,11 +4,5 @@ Contains robot interface, kinematics, and visualization components.
 """
 
 from .robot_interface import RobotInterface
-from .visualizer import PyBulletVisualizer
 
-__all__ = [
-    "RobotInterface",
-    "IKSolver",
-    "ForwardKinematics",
-    "PyBulletVisualizer",
-]
+__all__ = ["RobotInterface"]
