+++ telegrip/core/kinematics.py
@@ -198,14 +198,14 @@ class Arm_IK:
         opts = {"ipopt": {"print_level": 0, "max_iter": 50, "tol": 1e-4}, "print_time": False}
         self.opti.solver("ipopt", opts)
 
-    def ik_fun(self, target_pose, gripper=0, motorstate=None, motorV=None):
-        start_time = time.time()
+    def ik_fun(self, target_pose, gripper=0, motorstate=None, motorV=None, visualize=True):
         gripper = np.array([gripper / 2.0, -gripper / 2.0])
         if motorstate is not None:
             self.init_data = motorstate
         self.opti.set_initial(self.var_q, self.init_data)
 
-        # self.vis.viewer["ee_target"].set_transform(target_pose)  # for visualization
+        if visualize:
+            self.vis.viewer["ee_target"].set_transform(target_pose)  # for visualization
 
         self.opti.set_value(self.param_tf, target_pose)
         # self.opti.set_value(self.var_q_last, self.init_data) # for smooth
@@ -226,9 +226,8 @@ class Arm_IK:
                 self.init_data = sol_q
             self.history_data = sol_q
 
-            # self.vis.display(sol_q)  # for visualization
-            end_time = time.time()
-            print(f"IK mobile timer: {end_time - start_time} seconds")
+            if visualize:
+                self.vis.display(sol_q)  # for visualization
 
             if motorV is not None:
                 v = motorV * 0.0
