+++ telegrip/inputs/__init__.py
@@ -4,11 +4,9 @@ Contains VR WebSocket server and keyboard listener implementations.
 """
 
 from .base import ControlGoal
-from .keyboard_listener import KeyboardListener
 from .vr_ws_server import VRWebSocketServer
 
 __all__ = [
     "VRWebSocketServer",
-    "KeyboardListener",
     "ControlGoal",
 ]
